+++ src/frontend/apps/drive/src/features/explorer/utils/mimeTypes.ts
@@ -8,6 +8,7 @@ import mimePowerpoint from "@/assets/files/icons/mime-powerpoint.svg";
 import mimeAudio from "@/assets/files/icons/mime-audio.svg";
 import mimeVideo from "@/assets/files/icons/mime-video.svg";
 import mimeSQLite from "@/assets/files/icons/mime-sqlite.svg";
+import mimeGrist from "@/assets/files/icons/mime-grist.svg";
 
 import mimeCalcMini from "@/assets/files/icons/mime-calc-mini.svg";
 import mimeDocMini from "@/assets/files/icons/mime-doc-mini.svg";
@@ -19,6 +20,7 @@ import mimeAudioMini from "@/assets/files/icons/mime-audio-mini.svg";
 import mimeVideoMini from "@/assets/files/icons/mime-video-mini.svg";
 import mimeArchiveMini from "@/assets/files/icons/mime-archive-mini.svg";
 import mimeSQLiteMini from "@/assets/files/icons/mime-sqlite-mini.svg";
+import mimeGristMini from "@/assets/files/icons/mime-grist-mini.svg";
 
 
 import mimeArchive from "@/assets/files/icons/mime-archive.svg";
@@ -35,6 +37,7 @@ export enum MimeCategory {
   VIDEO = "video",
   ARCHIVE = "archive",
   SQLITE = "sqlite",
+  GRIST = "grist",
 }
 
 export const MIME_TO_ICON = {
@@ -48,6 +51,7 @@ export const MIME_TO_ICON = {
   [MimeCategory.VIDEO]: mimeVideo,
   [MimeCategory.ARCHIVE]: mimeArchive,
   [MimeCategory.SQLITE]: mimeSQLite,
+  [MimeCategory.GRIST]: mimeGrist,
 };
 
 export const MIME_TO_ICON_MINI = {
@@ -61,6 +65,7 @@ export const MIME_TO_ICON_MINI = {
   [MimeCategory.VIDEO]: mimeVideoMini,
   [MimeCategory.ARCHIVE]: mimeArchiveMini,
   [MimeCategory.SQLITE]: mimeSQLiteMini,
+  [MimeCategory.GRIST]: mimeGristMini,
 };
 
 export const ICONS = {
@@ -75,6 +80,7 @@ export const ICONS = {
     [MimeCategory.VIDEO]: mimeVideoMini,
     [MimeCategory.ARCHIVE]: mimeArchiveMini,
     [MimeCategory.SQLITE]: mimeSQLiteMini,
+    [MimeCategory.GRIST]: mimeGristMini,
 	},
 	"normal": {
     [MimeCategory.CALC]: mimeCalc,
@@ -87,6 +93,7 @@ export const ICONS = {
     [MimeCategory.VIDEO]: mimeVideo,
     [MimeCategory.ARCHIVE]: mimeArchive,
     [MimeCategory.SQLITE]: mimeSQLite,
+    [MimeCategory.GRIST]: mimeGrist,
 	}
 }
 
@@ -102,6 +109,7 @@ export const MIME_TO_FORMAT_TRANSLATION_KEY = {
   [MimeCategory.VIDEO]: "mime.video",
   [MimeCategory.ARCHIVE]: "mime.archive",
   [MimeCategory.SQLITE]: "mime.sqlite",
+  [MimeCategory.GRIST]: "mime.grist",
 };
 
 
@@ -152,6 +160,10 @@ export const getMimeCategory = (mimetype: string, extension?: string | null): Mi
       return MimeCategory.CALC;
   }
 
+  // Special case: a SQLITE file that has a .grist extension is a grist file
+  if (mimetype === "application/vnd.sqlite3" && extension && extension === "grist") {
+      return MimeCategory.GRIST;
+  }
 
   
   if (MIME_TO_CATEGORY[mimetype]) {